requirements.runpod.txt
# RunPod specific requirements for ACE-Step serverless worker

# RunPod SDK
runpod>=1.6.2

# Core ACE-Step dependencies
datasets==3.4.1
diffusers>=0.33.0
librosa==0.11.0
loguru==0.7.3
matplotlib==3.10.1
numpy
pypinyin==0.53.0
pytorch_lightning==2.5.1
soundfile==0.13.1
torchaudio
torchvision
tqdm
transformers==4.50.0
py3langid==0.3.0
hangul-romanize==0.1.0
num2words==0.5.14
spacy==3.8.4
accelerate==1.6.0
cutlet
fugashi[unidic-lite]
click
peft
tensorboard
tensorboardX

# Additional dependencies for enhanced functionality
hf_transfer
safetensors
scipy
Pillow